Changed ints back to fast_intN's. Also started to add documentation.

+++ Game/Characters/Character.cpp
@@ -195,7 +195,7 @@ sf::FloatRect Character::getBounds() const {
 
 /// Get curent experience points.
 /// \return Returns the current experience points of the Character.
-int Character::getExperience() const{
+int_fast16_t Character::getExperience() const{
 	return experiencePoints;
 }
 
@@ -203,14 +203,14 @@ int Character::getExperience() const{
 /// Set player experience points.
 /// \details
 /// This function adds the given experience points to the current amount of Character experience.
-void Character::setExperience(const int & experiencePointsToAdd){
+void Character::setExperience(const int_fast16_t & experiencePointsToAdd){
 	experiencePoints += experiencePointsToAdd;
 }
 
 /// \brief
 /// Get player health points.
 /// \return Returns the current health points of the Character.
-int Character::getHealth() const{
+int_fast8_t Character::getHealth() const{
 	return health;
 }
 
@@ -218,7 +218,7 @@ int Character::getHealth() const{
 /// Set player health points.
 /// \details
 /// This function adds the given health points to the current amount of Character health.
-void Character::setHealth(const int newHealth){
+void Character::setHealth(const int_fast8_t newHealth){
 	health = newHealth;
 	std::cout << "Health: " << int(health) << std::endl;
 }
